script.sh edited online with Bitbucket

+++ script.sh
@@ -13,8 +13,11 @@ clear
 echo "Installing g++"
 sudo apt-get install g++
 clear
+if ! [ -f "cuda_10.0.130_410.48_linux.run" ]
+then
 echo "Download CUDA drivers"
 wget https://developer.nvidia.com/compute/cuda/10.0/Prod/local_installers/cuda_10.0.130_410.48_linux -O cuda_10.0.130_410.48_linux.run
 clear
+fi
 echo "Initiating CUDA driver installation"
 sudo sh cuda_10.0.130_410.48_linux.run
